lunch/lunch/requirements.txt
Django==1.8
django-braces==1.4.0
django-crispy-forms-ng==2.0.0
git+https://github.com/django-tastypie/django-tastypie.git
django-toolbelt
djangorestframework
six==1.9.0
